backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
django-filter==23.5
openpyxl==3.1.2
xhtml2pdf==0.2.11
weasyprint==66.0
PyPDF2==3.0.1

